# docs/swagger.yaml
basePath: /
definitions:
  apierr.APIError:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
    type: object
...
  topics.CreateTopicRequest:
    properties:
      category:
        maxLength: 50
        minLength: 1
        type: string
      name:
        maxLength: 150
        minLength: 1
        type: string
      slug:
        maxLength: 100
        minLength: 2
        type: string
    required:
    - category
    - name
    - slug
    type: object
  topics.TopicResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      is_system:
        type: boolean
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  topics.UpdateTopicRequest:
    properties:
      category:
        maxLength: 50
        minLength: 1
        type: string
      name:
        maxLength: 150
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: felipe@example.com
    name: Felipe
  description: REST API con Go + Gin · Auth JWT · PostgreSQL · Swagger
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: dev-test-api
  version: "1.0"
paths:
...
  /api/v1/topics:
    get:
      description: Lista los temas según el rol del usuario. Admin ve solo temas del
        sistema (is_system=true). Usuarios normales ven solo sus propios temas personalizados
        (is_system=false).
      parameters:
      - description: 'Número de página (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Elementos por página (default: 20, max: 100)'
        in: query
        name: per_page
        type: integer
      - description: 'Campo de ordenación: name, slug, category, created_at'
        in: query
        name: sort_by
        type: string
      - description: 'Dirección: asc o desc (default: desc)'
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de temas (con paginación)
          schema:
            $ref: '#/definitions/response.Meta'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Listar temas (con paginación)
      tags:
      - topics
    post:
      consumes:
      - application/json
      description: Crea un nuevo tema. Admin crea temas del sistema (is_system=true),
        usuarios normales crean temas personalizados (is_system=false).
      parameters:
      - description: Datos del tema
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/topics.CreateTopicRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/topics.TopicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Crear tema
      tags:
      - topics
  /api/v1/topics/{id}:
    delete:
      description: Elimina un tema. Admin solo puede eliminar temas del sistema (is_system=true).
        Usuarios normales solo pueden eliminar sus propios temas (is_system=false).
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Eliminar tema
      tags:
      - topics
    get:
      description: Obtiene un tema por ID. Admin solo puede ver temas del sistema
        (is_system=true). Usuarios normales pueden ver temas del sistema y sus propios
        temas.
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/topics.TopicResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Obtener tema
      tags:
      - topics
    put:
      consumes:
      - application/json
      description: Actualiza un tema existente. Admin solo puede modificar temas del
        sistema (is_system=true). Usuarios normales solo pueden modificar sus propios
        temas (is_system=false).
      parameters:
      - description: Topic ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/topics.UpdateTopicRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/topics.TopicResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierr.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apierr.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apierr.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/apierr.APIError'
      security:
      - BearerAuth: []
      summary: Actualizar tema
      tags:
      - topics
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'JWT token. Prefix with "Bearer ". Example: "Bearer eyJhbG..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"